+++ BxInputController/Sources/Sections/BxInputSection.swift
@@ -16,10 +16,26 @@ import UIKit
 /// Section class with all content data for BxInputController
 open class BxInputSection
 {
-
-    /// metadata of cells for this section
-    open var rowBinders: [BxInputRowBinder] = []
     
+    /// rows content without metadata
+    open var rows: [BxInputRow]
+    {
+        get {
+            var result: [BxInputRow] = []
+            for rowBinder in rowBinders {
+                result.append(rowBinder.rowData)
+            }
+            return result
+        }
+        set {
+            rowBinders = []
+            for row in newValue
+            {
+                rowBinders.append(row.binder)
+            }
+        }
+    }
+
     /// header content of the section, that showed above rows
     open var header: BxInputSectionContent? = nil
     {
@@ -50,13 +66,12 @@ open class BxInputSection
     internal(set) public var headerBinder: BxInputSectionContentBinder? = nil
     /// binder with logic of working footer view and data model
     internal(set) public var footerBinder: BxInputSectionContentBinder? = nil
+    /// binder with logic of working cells for this section
+    internal(set) public var rowBinders: [BxInputRowBinder] = []
     
     /// construction object with text header and/or fother, default headerText/footerText is nil and they won't show
     public init(headerText: String? = nil, rows: [BxInputRow], footerText: String? = nil) {
-        for row in rows
-        {
-            rowBinders.append(row.binder)
-        }
+        self.rows = rows
         defer {
             if let headerText = headerText {
                 header = BxInputSectionStringHeader(headerText)